Adds a polyphase FIR decimation filter to digitalDemod.
This adds an anti-aliasing filter to the decimation step referenced in
b55a801d5c63ee3c239f313ccac054fca38eef0b. The final filter has as many
as three stages of decimation, depending on the choice of IF frequency
and channel bandwidth.

Testing on MQCO11 shows that with: 10 MHz IF, 5 MHz bandwidth, and 500
MHz sampling rate, the resulting filter can be applied to a 2K point
buffer in 10 us.

--BRJ and CAR